requirements.txt
torch
torchvision
torchaudio
numpy
pyyaml
pandas
tqdm
scikit-learn
matplotlib
torchmetrics
opencv-python
torchxrayvision